feat: 增加demo配置和结构
