+++ CMakeLists.txt
@@ -11,7 +11,7 @@ endif()
 
 # 安装路径
 if (NOT INSTALL_PREFIX)
-    set(INSTALL_PREFIX "${CMAKE_CURRENT_LIST_DIR}/bin/${CMAKE_SYSTEM_NAME}")
+    set(INSTALL_PREFIX "${CMAKE_CURRENT_BINARY_DIR}/bin/${CMAKE_SYSTEM_NAME}")
 elseif(${INSTALL_PREFIX} STREQUAL "." OR ${INSTALL_PREFIX} STREQUAL "./")
     set(INSTALL_PREFIX "${CMAKE_CURRENT_BINARY_DIR}/bin/${CMAKE_SYSTEM_NAME}")
 else()
@@ -26,8 +26,10 @@ endif()
 # 指定编译器标准
 if(NOT DEFINED CXX_STANDARD)
     set(CMAKE_CXX_STANDARD 17)
-else()
+elseif (CXX_STANDARD EQUAL 11 OR CXX_STANDARD EQUAL 14 OR CXX_STANDARD EQUAL 17)
     set(CMAKE_CXX_STANDARD ${CXX_STANDARD})
+else()
+    message(FATAL_ERROR "CXX_STANDARD(${CXX_STANDARD}) not supported")
 endif()
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
 
@@ -37,13 +39,14 @@ if (NOT CMAKE_BUILD_TYPE)
     set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -g")
     add_definitions(-DDEBUG)
 else()
-    if(${CMAKE_BUILD_TYPE} STREQUAL "release")
+    string(TOLOWER ${CMAKE_BUILD_TYPE} BUILD_TYPE)
+    if(${BUILD_TYPE} STREQUAL "release")
         set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -O2")
-    elseif(${CMAKE_BUILD_TYPE} STREQUAL "debug")
+    elseif(${BUILD_TYPE} STREQUAL "debug")
         set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -g")    # -g
         add_definitions(-DDEBUG)                        # 增加debug宏，开启源码中的DEBUG能力
     else()          # 没有指定就使用debug模式
-        message(FATAL_ERROR "CMAKE_BUILD_TYPE not supported")
+        message(FATAL_ERROR "CMAKE_BUILD_TYPE(${CMAKE_BUILD_TYPE}) not supported")
     endif()
 endif()
 
@@ -68,7 +71,7 @@ set(APP_PATH "${CMAKE_CURRENT_LIST_DIR}/app")
 include_directories("${APP_PATH}/src")
 include_directories("${APP_PATH}/include")
 set(APP_OUTPUT "${INSTALL_PREFIX}")
-set(APP_NAME "MeowCut")
+set(APP_TARGET "MeowCut")
 
     # core
 set(CORE_PATH "${CMAKE_CURRENT_LIST_DIR}/core")
@@ -76,18 +79,44 @@ include_directories("${CORE_PATH}/src")
 include_directories("${CORE_PATH}/third_party")
 include_directories("${CORE_PATH}/include")
 set(CORE_OUTPUT "${INSTALL_PREFIX}")
-set(CORE_TARGET "MeowCutCore")
+set(CORE_TARGET "${APP_TARGET}Core")
 
 if (NOT DEFINED SOFTWARE_TYPE)
     add_subdirectory("${CORE_PATH}")
     add_subdirectory("${APP_PATH}")
 else()
-    if ("${SOFTWARE_TYPE}" STREQUAL "lib")      # build lib only
+    string(TOLOWER ${SOFTWARE_TYPE} _SOFTWARE_TYPE_)
+    if ("${_SOFTWARE_TYPE_}" STREQUAL "lib")      # build lib only
         add_subdirectory("${CORE_PATH}")
-    elseif("${SOFTWARE_TYPE}" STREQUAL "app")                                      # build lib and app
+    elseif("${_SOFTWARE_TYPE_}" STREQUAL "app")                                      # build lib and app
         add_subdirectory("${CORE_PATH}")
         add_subdirectory("${APP_PATH}")
+    elseif("${_SOFTWARE_TYPE_}" STREQUAL "demo")
+        # 设置测试的demo
+        if (NOT DEMO_TYPE)
+            message(FATAL_ERROR "DEMO_TYPE must set while build demo")
+        else()
+            string(TOLOWER ${DEMO_TYPE} _DEMO_TYPE_)
+            if ("${_DEMO_TYPE_}" STREQUAL "audio_effect")
+                add_definitions(-DDEMO_AUDIO_EFFECT)
+            elseif("${_DEMO_TYPE_}" STREQUAL "avcodec")
+                add_definitions(-DDEMO_AVCODEC)
+            elseif("${_DEMO_TYPE_}" STREQUAL "avsync")
+                add_definitions(-DDEMO_AVSYNC)
+            elseif("${_DEMO_TYPE_}" STREQUAL "image_filter")
+                add_definitions(-DDEMO_IMAGE_FILTER)
+            elseif("${_DEMO_TYPE_}" STREQUAL "meowcut_core")
+                add_definitions(-DDEMO_MEOWCUT_CORE)
+            elseif("${_DEMO_TYPE_}" STREQUAL "video_render")
+                add_definitions(-DDEMO_VIDEO_RENDER)
+            else()
+                message(FATAL_ERROR "DEMO_TYPE(${DEMO_TYPE}) not supported")
+            endif()
+        endif()
+        add_definitions(-DDEBUG)            # demo must in debug
+        # set(BUILD_DEMO ON)
+        add_subdirectory("${CORE_PATH}")
     else()
-        message(FATAL_ERROR "SOFTWARE_TYPE not suooprted")
+        message(FATAL_ERROR "SOFTWARE_TYPE(${SOFTWARE_TYPE}) not suooprted")
     endif()
 endif()
