+++ app/CMakeLists.txt
@@ -1,4 +1,4 @@
-project(${APP_NAME} VERSION ${SOFTWARE_VERSION} LANGUAGES CXX)
+project(${APP_TARGET} VERSION ${SOFTWARE_VERSION} LANGUAGES CXX)
 
 set(CMAKE_AUTOUIC ON)
 set(CMAKE_AUTOMOC ON)
