+++ core/src/avcodec/demo/avcodec.qrc
@@ -1,0 +1,4 @@
+<RCC>
+    <qresource prefix="AVCodecDemo">
+    </qresource>
+</RCC>
