--- core/src/avcodecwidget/avcodecwidget.qrc
@@ -1,4 +1,0 @@
-<RCC>
-    <qresource prefix="AVCodecWidget">
-    </qresource>
-</RCC>
