--- core/src/avcodecwidget/main.cpp
@@ -1,10 +1,0 @@
-#include "avcodecwidget.h"
-#include <QtWidgets/QApplication>
-
-int main(int argc, char *argv[])
-{
-    QApplication a(argc, argv);
-    AVCodecWidget w;
-    w.show();
-    return a.exec();
-}
